backend/google_auth/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
PyJWT==2.10.1
